Refactor of external and streaming catalog
