+++ core/src/main/scala/org/apache/spark/sql/crossdata/catalog/utils/CatalogUtils.scala
@@ -1,0 +1,51 @@
+/*
+ * Copyright (C) 2015 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.spark.sql.crossdata.catalog.utils
+
+import java.lang.reflect.Constructor
+
+import com.typesafe.config.Config
+import org.apache.spark.Logging
+import org.apache.spark.sql.catalyst.CatalystConf
+import org.apache.spark.sql.crossdata.XDContext
+import org.apache.spark.sql.crossdata.catalog.interfaces.{XDPersistentCatalog, XDStreamingCatalog}
+
+object CatalogUtils extends Logging {
+
+  protected[crossdata] def externalCatalog(catalystConf: CatalystConf, config: Config): XDPersistentCatalog = {
+    import XDContext.DerbyClass
+    val externalCatalogName = if (config.hasPath(XDContext.ClassConfigKey))
+      config.getString(XDContext.ClassConfigKey)
+    else DerbyClass
+
+    val externalCatalogClass = Class.forName(externalCatalogName)
+    val constr: Constructor[_] = externalCatalogClass.getConstructor(classOf[CatalystConf])
+
+    constr.newInstance(catalystConf).asInstanceOf[XDPersistentCatalog]
+  }
+
+  protected[crossdata] def streamingCatalog(catalystConf: CatalystConf, config: Config): Option[XDStreamingCatalog] = {
+    if (config.hasPath(XDContext.StreamingCatalogClassConfigKey)) {
+      val streamingCatalogClass = config.getString(XDContext.StreamingCatalogClassConfigKey)
+      val xdStreamingCatalog = Class.forName(streamingCatalogClass)
+      val constr: Constructor[_] = xdStreamingCatalog.getConstructor(classOf[CatalystConf])
+      Option(constr.newInstance(catalystConf).asInstanceOf[XDStreamingCatalog])
+    } else {
+      logError("Empty streaming catalog")
+      None
+    }
+  }
+}
